+++ packages/definitions/CHANGELOG.md
@@ -1,3 +1,8 @@
+**0.0.4**
+
+* rename `LocoKitField` to `LocoKitFormField`
+* rename `LocoKitFieldRule` to `LocoKitFormFieldRule`
+
 **0.0.3**
 
-* add LocoKitField and LocoKitFieldRule types
+* add `LocoKitField` and `LocoKitFieldRule` types